save deskewed data as uint16

+++ mantis/cli/deskew.py
@@ -125,7 +125,7 @@ def deskew_zyx_and_save(
     )
     # Write to file
     with open_ome_zarr(output_path, mode="r+") as output_dataset:
-        output_dataset[0][t_idx, c_idx] = deskewed
+        output_dataset[0][t_idx, c_idx] = np.uint16(deskewed)
         output_dataset.zattrs["deskewing"] = asdict(settings)
 
     click.echo(f"Finished Writing.. c={c_idx}, t={t_idx}")
